End-points to manage users and databases on the server were added. (#582)
Public API:

This PR adds a public API to work with databases on the server using
`YouTracks#instance(host, port, user, password)` method to connect to
the server.

After that, the user can add server-level users to the server using the
YouTrackDB#createSystemUser method. The following PRs will add more
methods to handle server user methods.

Internal implementation:
1. `YTDBGremlinSaslAuthenticationHandler` was added to send dbName to
the server as authorizationId. That will allow future PRs to perform
user queries to the server, not on behalf of server-wide (system users),
but on behalf of a user registered in a single database.
2. `YTDBDriverRemoteConnection` and `YTDBDriverRemoteTraversal` are now
tracking the changes of RIDs of new records instead of
`YTDBGremlinResponseHandler` that in new implementation only passes this
mapping to the respective classes. That allows us to update RIDs of
committed vertices not only at the level of a single query but on the
level of several queries that are executed inside a single transaction.
3. Tests for handling of TX on the server level were added, though we
need to have more tests for TX tracking in the future.
4. YTDB-specific OpProcessors were added. They are, in a nutshell, a
reflection of the original implementation with two important
differences:
- We use only ANTL for parsing of Gremlin-Script, so no Groovy anymore,
as it is a big security hole.
- We work with only a single database on a single traversal level, which
allows us to integrate Gremlin Server lifecycle with our own lifecycle
seamlessly.