Two GitHub action workflows were added. (#589)
The first workflow runs every time a PR is created, added to the merge
queue, or merged into the develop branch.
When running this workflow in the develop branch, build artifacts are
deployed to Maven Central, both as SNAPSHOT versions and as SNAPSHOT
versions with a short commit SHA suffix.
That allows users to work with a fixed SNAPSHOT version.
The first workflow utilizes disk-backed storage for Ubuntu x86 builds
and memory-backed storage for ARM and Windows builds.
The second workflow runs periodically, once a day, in case new changes
are submitted after the last successful run of this workflow.
The second workflow uses the disk-backed storage for all types of
databases. In all workflows, failures are reported to Zulip chat. In all
workflows, tests run on Windows, Ubuntu x86, and Ubuntu ARM nodes on JDK
21 and JDK 25 in parallel.

---------

Co-authored-by: Vladislav <39089602+lesley29@users.noreply.github.com>